requirements.txt
Flask==3.0.0
pdfplumber==0.11.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
requests==2.31.0
